requirements.txt
flask==3.0.3
pymysql==1.1.1
paddlepaddle==2.6.2
paddleocr==2.7.3
pillow==10.3.0
numpy==1.26.4
werkzeug==3.0.3

